chore: reorganiza pacotes
