+++ apps/api/src/main/java/br/org/apae/api/profissional/da/saude/domain/exceptions/DadosInvalidosException.java
@@ -1,4 +1,4 @@
-package br.org.apae.api.profissional.da.saude.exceptions;
+package br.org.apae.api.profissional.da.saude.domain.exceptions;
 
 public class DadosInvalidosException extends RuntimeException{
     public DadosInvalidosException(String message) {
