+++ apps/api/src/main/java/br/org/apae/api/profissional/da/saude/domain/exceptions/EntidadeNaoEncontradaException.java
@@ -1,4 +1,4 @@
-package br.org.apae.api.profissional.da.saude.exceptions;
+package br.org.apae.api.profissional.da.saude.domain.exceptions;
 
 public class EntidadeNaoEncontradaException extends RuntimeException{
     public EntidadeNaoEncontradaException(String message) {
