+++ apps/api/src/main/java/br/org/apae/api/profissional/da/saude/domain/repository/ProfissionalSaudeRepository.java
@@ -1,10 +1,6 @@
 package br.org.apae.api.profissional.da.saude.domain.repository;
 
-import br.org.apae.api.profissional.da.saude.domain.model.ProfissionalSaude;
-import br.org.apae.api.profissional.da.saude.entity.ProfissionalSaudeEntity;
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
+import br.org.apae.api.profissional.da.saude.infra.entity.ProfissionalSaudeEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
